+++ state/app/types.ts
@@ -1,8 +1,8 @@
 import { BigNumber } from 'ethers';
 
-import { KwentaStatus } from 'sdk/services/system';
 import { TransactionStatus } from 'sdk/types/common';
 import { FuturesMarketKey } from 'sdk/types/futures';
+import { KwentaStatus } from 'sdk/types/system';
 import { FuturesTransactionType } from 'state/futures/types';
 
 export type ModalType =
